Fix savepoint leak on nested transaction rollback
Prior to this commit rollback of nested transaction didn't released
underlying savepoint but cleared it AbstractTransactionStatus thus
savepoint was not released until end of outer transaction.

This fix releases savepoint on rollback. Though rollback to savepoint
in databases does allow to use given savepoint again the lifecycle of
spring managed transaction which is tracked by
AbstractTransactionStatus does not allow for transaction to procceed
after rollback. That is transaction is considered completed if either
rollback or commit is performed. So further reuse of the underlying
savepoint is not needed and it should be cleared on rollback.

Issue: SPR-12228